pipeline/requirements.txt
aiohttp==3.13.2
apache-airflow-providers-ssh==3.12.0
apache-airflow==2.10.3
asyncpg==0.29.0
beautifulsoup4==4.12.3
chain-index==0.3.0
dune-client==1.7.4
eth-typing==5.2.1
eth-utils==5.3.1
fastapi==0.111.0
hexbytes==1.3.1
httpx==0.27.2
hvac==2.1.0
igraph==0.11.3
loguru==0.7.2
niquests==3.5.5
openrank-sdk==0.2.2
pandas==2.1.3
polars==0.20.27
psutil==5.9.8
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic-settings==2.2.1
pydantic==2.11.7
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
sqlalchemy==1.4.52
supabase==2.9.0
tldextract==5.1.1
tomlkit==0.13.2
urllib3==2.5.0
web3==7.14.0
